Added functionality for block manager, added temp to gitignore
BlockManager will now handle a blocks attached effects. To add an effect, add the desired effect's script to the block prefab. Supports modes for triggering on hit, on damaged and both.

Also added the Temp/ subfolder to gitignore just in case somebody wants to commit while Unity is still running.