Added Revival Block Behaviour as well as Target Area Mechanic
The target manager will assemble all of the targetarea blocks at the start of a level and mark the level complete when all targetareas are active at once. Target areas deactivate themselves after a predetermined amount of time or ball bounces is met if the level isn't completed before then.

Also added the revival block behaviour which can be set to make blocks respawn after a certain amount of time or ball bounces.